requirements.txt
streamlit
pandas
openai
langchain
matplotlib
plotly
openpyxl
python-dotenv